requirements.txt
matplotlib==3.8.2
numpy==1.24.1
pandas==2.2.2
python-dotenv==1.0.1
scikit_learn==1.4.1.post1
scikit_learn==1.3.2
torch==2.1.1+cu121
